refactor: 아이디 공유하기 컴포넌트 분리
